requirements/external_apps.txt
# Get all requirements formerly known as "external_libs" explicitly
-r libs.txt

# Already released apps
django-ajax-validation
django-announcements
django-avatar
django-dbtemplates
django-microblogging
django-notification
django-pagination
django-robots
django-messages
django-imagekit
django-wikiapp==0.1.1

# Apps that are installed "directly" non-editable because they lack a
# setuptools setup.py, are not released on pypi or don't have a recent release
http://django-tagging.googlecode.com/svn/trunk/
http://django-voting.googlecode.com/svn/trunk/
http://django-timezones.googlecode.com/svn/trunk/
http://django-command-extensions.googlecode.com/svn/trunk/
http://django-db-log.googlecode.com/svn/trunk/
http://django-page-cms.googlecode.com/svn/trunk/
http://django-mptt.googlecode.com/svn/trunk/

# Do editable checkouts (yay!)
-e svn+http://django-bookmarks.googlecode.com/svn/trunk#egg=django-bookmarks
-e svn+http://django-email-confirmation.googlecode.com/svn/trunk#egg=django-email-confirmation
-e svn+http://django-flag.googlecode.com/svn/trunk#egg=django-flag
-e svn+http://django-friends.googlecode.com/svn/trunk#egg=django-friends
-e svn+http://django-gravatar.googlecode.com/svn/trunk/#egg=django-gravatar
-e svn+http://django-mailer.googlecode.com/svn/trunk/#egg=django-mailer
-e svn+http://django-oembed.googlecode.com/svn/trunk/#egg=django-oembed
-e svn+http://django-schedule.googlecode.com/svn/trunk#egg=django-schedule
-e svn+http://django-threadedcomments.googlecode.com/svn/trunk#egg=django-threadedcomments
-e svn+http://django-locations.googlecode.com/svn/trunk#egg=django-locations
-e svn+http://django-uni-form.googlecode.com/svn/trunk#egg=django-uni-form
-e svn+http://django-swaps.googlecode.com/svn/trunk#egg=django-swaps
-e svn+http://django-app-plugins.googlecode.com/svn/trunk#egg=django-app-plugins
-e git://github.com/brosner/django-openid.git#egg=django-openid
-e git://github.com/robhudson/django-debug-toolbar.git#egg=django-debug-toolbar
-e git://github.com/jezdez/django-feedutil.git#egg=django-feedutil
